Update doc with location of the component
Love this new component! At first, I added it to App.svelte because I wanted to customize the color. I ended up with two headers on my page since it's already being used in index.svelte. It took me a minute to figure out where it was already being imported. My thought is that by adding the location of the component to the docs, we can avoid the confusion I had.